2005-11-02  Roman Kennke  <kennke@aicas.com>
	* java/awt/Component.java
	(getMaximumSize): Return (Short.MAX_VALUE, Short.MAX_VALUE) instead
	of (Integer.MAX_VALUE, Integer.MAX_VALUE) as shown by the
	Mauve test.

+++ java/awt/Component.java
@@ -1645,7 +1645,7 @@ public abstract class Component
    */
   public Dimension getMaximumSize()
   {
-    return new Dimension(Integer.MAX_VALUE, Integer.MAX_VALUE);
+    return new Dimension(Short.MAX_VALUE, Short.MAX_VALUE);
   }
 
   /**
